// src/server/interfaces/iUser.interface.ts
export interface iUser {
  id: number;
  firstName: string;
  lastName: string;
  companyId: number;
  userName: string;
  password: string;
  roleId: number;
  emailAddress: string;
  markupId: number;
  resetPasswordExpires: string;
  resetPasswordToken: string;
  salt: string;
  tAdmin_Role: any;
}